requirements.txt
Flask==2.0.0
line-bot-sdk==1.19.0
requests==2.25.1
beautifulsoup4 = "*"
lxml = "*"
numpy = "*"